+++ railties/lib/rails/generators/rails/plugin/plugin_generator.rb
@@ -86,7 +86,7 @@ task default: :test
     end
 
     PASSTHROUGH_OPTIONS = [
-      :skip_active_record, :skip_action_mailer, :skip_javascript, :skip_sprockets, :database,
+      :skip_active_record, :skip_action_mailer, :skip_javascript, :skip_action_cable, :skip_sprockets, :database,
       :javascript, :quiet, :pretend, :force, :skip
     ]